Current version of XbeeCom
This code sends dummy grid eye data (128 bytes) along with a sensor unit identifier, data type and delimeter. This is the format of the packet it sends.

17 bytes
These bytes are used by the xBee in API mode.
0x7E, 0x00, 0x92, 0x10, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xFF, 0xFE, 0x00, 0x00

1 byte
Sensor unit identifier. right now it's just 1, but as we add more units it will go higher

1 byte
Data Type identifier. 1 is for grid-eye data. other numbers will be used for other data types such as CO2, humidity, and temperature.

128 bytes. 
Grid eye data. Note that each temperature 'pixel' has a high byte and a low byte. These will have to be combined while parsing to get the actual temperature.

2 bytes.
Delimeter. 0xDF, 0xDF. I put these two bytes here so that when you parse the data you can know when you've reached the end.

1 byte.
Checksum. This is used by the xBee in API mode to check the data. You can also make use of it if you want.
To ve