Added support for RHT03 Temp and Humidity Sensor
The unit now reads the temperature and humidity from the RHT03. The data from the sensor is 4 bytes long and is stored in the Xbee packet between the duty cycle and the grideye data. The first two bytes correspond with the humidity and the second two bytes are for temperature. Concatenate the first two and divide the result by 10 to get the humidity as a percentage and do the same to get the temperature in Celsius.